Styling and naming tweaks to renderer files
